win_requirements.txt
anthropic
backoff
numpy
openai
playwright
# pygraphviz
pynput
pytest
python-dotenv
requests
rich
sounddevice
soundfile
typing-extensions
